DataPipeliner/requirements.txt
pandas==2.2.2
pyarrow==21.0.0
duckdb==1.3.2
deltalake==1.1.3
beautifulsoup4==4.12.3
minio==7.2.0
python-dotenv==1.1.1
langdetect==1.0.9
textstat==0.7.8
numpy==1.26.4
sentence-transformers==5.0.0
chromadb==1.0.15
requests==2.32.4
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
ollama==0.5.1
apache-airflow